Update RA3p1
updated. created output files for both mtDNA and Ychrom. Then started off with openin mtDNA file,  parsing it into a dict, then naming each sequence by calling them based on IDs(i.e. the names). Then iterated over all dictionary values over the second sequence (took this as an example). Was thinking of doing this for the rest as well but was also thinking of trying to apply more for loops